+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - pinned conda and mamba versions in `.travis.yaml` 
   - temp fix until conda>=4.12 can install mamba properly
 - documentation is working again!
+- code in gimmemotifs.report now supports pandas >=1.30
 
 ### Removed